(refactor) Splits bindings into submodules and adds LocalMapperGPU bindings
