+++ src/kompass_cpp/kompass_cpp/include/mapping/local_mapper_gpu.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#ifndef GPU
+#define GPU 1
+#endif // !GPU
+
 #include "local_mapper.h"
 #include <Eigen/Dense>
 #include <sycl/sycl.hpp>
