+++ src/kompass_cpp/kompass_cpp/src/bindings/bindings_control.cpp
@@ -2,204 +2,20 @@
 #include <pybind11/functional.h>
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
-#include <vector>
 
-#include "controllers/controller.h"
-#include "controllers/dwa.h"
+#include "controllers/follower.h"
+#include "datatypes/control.h"
+#include "datatypes/trajectory.h"
+#include "controllers/vision_follower.h"
 #include "controllers/pid.h"
 #include "controllers/stanley.h"
-#include "controllers/vision_follower.h"
-#include "datatypes/control.h"
-#include "datatypes/parameter.h"
-#include "datatypes/path.h"
-#include "datatypes/trajectory.h"
-#include "mapping/local_mapper.h"
-#include "utils/logger.h"
+#include "controllers/dwa.h"
 
 namespace py = pybind11;
-
 using namespace Kompass;
 
-// Define a variant type to hold different parameter value types
-using ParamValue = std::variant<double, int, bool, std::string>;
-
-std::string printControlCmd(const Control::Velocity &velocity_command) {
-  return "{" + std::to_string(velocity_command.vx) + ", " +
-         std::to_string(velocity_command.vy) + ", " +
-         std::to_string(velocity_command.omega) + "})";
-}
-
-// Method to set parameter values based on dict instance
-void set_parameters_from_dict(Parameters &params,
-                              const py::object &attrs_instance) {
-  auto attrs_dict = attrs_instance.cast<py::dict>();
-  for (const auto &item : attrs_dict) {
-    const std::string &name = py::str(item.first);
-    py::handle value = item.second;
-    try {
-      auto it = params.parameters.find(name);
-      if (it != params.parameters.end()) {
-        if (py::isinstance<py::bool_>(value)) {
-          it->second.setValue(value.cast<bool>());
-        } else if (py::isinstance<py::float_>(value)) {
-          it->second.setValue(value.cast<double>());
-        } else if (py::isinstance<py::str>(value)) {
-          it->second.setValue(py::str(value).cast<std::string>());
-        } else if (py::isinstance<py::int_>(value)) {
-          it->second.setValue(value.cast<int>());
-        }
-      }
-    } catch (const std::exception &e) {
-      PyErr_SetString(PyExc_RuntimeError, e.what());
-      throw py::error_already_set();
-    }
-  }
-}
-
-PYBIND11_MODULE(kompass_cpp, m) {
-  m.doc() = "Algorithms for robot path tracking and control";
-
-  // Types submodule
-  auto m_types = m.def_submodule("types", "KOMPASS CPP data types module");
-
-  py::enum_<Path::InterpolationType>(m_types, "PathInterpolationType")
-      .value("LINEAR", Path::InterpolationType::LINEAR)
-      .value("CUBIC_SPLINE", Path::InterpolationType::CUBIC_SPLINE)
-      .value("HERMITE_SPLINE", Path::InterpolationType::HERMITE_SPLINE);
-
-  py::class_<Path::State>(m_types, "State")
-      .def(py::init<double, double, double, double>(), py::arg("x") = 0.0,
-           py::arg("y") = 0.0, py::arg("yaw") = 0.0, py::arg("speed") = 0.0)
-      .def_readwrite("x", &Path::State::x)
-      .def_readwrite("y", &Path::State::y)
-      .def_readwrite("yaw", &Path::State::yaw)
-      .def_readwrite("speed", &Path::State::speed);
-
-  py::class_<Path::Point>(m_types, "Point")
-      .def(py::init<double, double>(), py::arg("x") = 0.0, py::arg("y") = 0.0)
-      .def_readwrite("x", &Path::Point::x)
-      .def_readwrite("y", &Path::Point::y);
-
-  py::class_<Path::PathPosition>(m_types, "PathPosition")
-      .def(py::init<>())
-      .def_readwrite("segment_index", &Path::PathPosition::segment_index)
-      .def_readwrite("segment_length", &Path::PathPosition::segment_length)
-      .def_readwrite("parallel_distance",
-                     &Path::PathPosition::parallel_distance)
-      .def_readwrite("normal_distance", &Path::PathPosition::normal_distance);
-
-  py::class_<Path::Path>(m_types, "Path")
-      .def(py::init<const std::vector<Path::Point> &>(),
-           py::arg("points") = std::vector<Path::Point>())
-      .def("reached_end", &Path::Path::endReached)
-      .def("get_total_length", &Path::Path::totalPathLength)
-      .def_readwrite("points", &Path::Path::points);
-
-  // Velocity control command
-  py::class_<Control::Velocity>(m_types, "ControlCmd")
-      .def(py::init<double, double, double, double>(), py::arg("vx") = 0.0,
-           py::arg("vy") = 0.0, py::arg("omega") = 0.0,
-           py::arg("steer_ang") = 0.0)
-      .def_readwrite("vx", &Control::Velocity::vx)
-      .def_readwrite("vy", &Control::Velocity::vy)
-      .def_readwrite("omega", &Control::Velocity::omega)
-      .def_readwrite("steer_ang", &Control::Velocity::steer_ang)
-      .def("__str__", &printControlCmd);
-
-  py::class_<Control::Velocities>(m_types, "ControlCmdList")
-      .def(py::init<>(), "Default constructor")
-      .def(py::init<int>(), "Constructor with length", py::arg("length"))
-      .def_readwrite("vx", &Control::Velocities::vx, "Speed on x-axis (m/s)")
-      .def_readwrite("vy", &Control::Velocities::vy, "Speed on y-axis (m/s)")
-      .def_readwrite("omega", &Control::Velocities::omega,
-                     "Angular velocity (rad/s)")
-      .def_readwrite("length", &Control::Velocities::_length,
-                     "Length of the vectors");
-
-  py::class_<Control::Trajectory>(m_types, "Trajectory")
-      .def(py::init<>())
-      .def_readwrite("velocity", &Control::Trajectory::velocity)
-      .def_readwrite("path", &Control::Trajectory::path);
-
-  py::class_<Control::LaserScan>(m_types, "LaserScan")
-      .def(py::init<std::vector<double>, std::vector<double>>(),
-           py::arg("ranges"), py::arg("angles"))
-      .def_readwrite("ranges", &Control::LaserScan::ranges)
-      .def_readwrite("angles", &Control::LaserScan::angles);
-
-  py::class_<Control::Point3D>(m_types, "Point3D")
-      .def(py::init<double, double, double>(), py::arg("x") = 0.0,
-           py::arg("y") = 0.0, py::arg("z") = 0.0)
-      .def_readwrite("x", &Control::Point3D::x)
-      .def_readwrite("y", &Control::Point3D::y)
-      .def_readwrite("z", &Control::Point3D::z);
-
-  // For collisions detection
-  py::enum_<CollisionChecker::ShapeType>(m_types, "RobotGeometry")
-      .value("CYLINDER", CollisionChecker::ShapeType::CYLINDER)
-      .value("BOX", CollisionChecker::ShapeType::BOX)
-      .def("get", [](const std::string &key) {
-        if (key == "CYLINDER")
-          return CollisionChecker::ShapeType::CYLINDER;
-        if (key == "BOX")
-          return CollisionChecker::ShapeType::BOX;
-        throw std::runtime_error("Invalid key");
-      });
-
-  // Vision types
-  py::class_<Control::VisionFollower::TrackingData>(m_types, "TrackingData")
-      .def(py::init<std::array<double, 2>, int, int, std::array<double, 2>,
-                    double>(),
-           py::arg("size_xy"), py::arg("img_width"), py::arg("img_height"),
-           py::arg("center_xy"), py::arg("depth") = -1.0)
-      .def_readwrite("size_xy", &Control::VisionFollower::TrackingData::size_xy)
-      .def_readwrite("img_width",
-                     &Control::VisionFollower::TrackingData::img_width)
-      .def_readwrite("img_height",
-                     &Control::VisionFollower::TrackingData::img_height)
-      .def_readwrite("center_xy",
-                     &Control::VisionFollower::TrackingData::center_xy)
-      .def_readwrite("depth", &Control::VisionFollower::TrackingData::depth);
-
-  // --------------------------------------------------------------------------------
-  // Config parameters bindings submodule
-  auto m_config = m.def_submodule("configure", "Configuration classes");
-
-  py::class_<Parameter>(m_config, "ConfigParameter")
-      .def(py::init<>())
-      .def(py::init<int, int, int>())
-      .def(py::init<double, double, double>())
-      .def(py::init<std::string>())
-      .def(py::init<bool>())
-      .def("set_value", (void(Parameter::*)(int)) & Parameter::setValue<int>)
-      .def("set_value",
-           (void(Parameter::*)(double)) & Parameter::setValue<double>)
-      .def("set_value",
-           (void(Parameter::*)(std::string)) & Parameter::setValue<std::string>)
-      .def("get_value_int", &Parameter::getValue<int>)
-      .def("get_value_double", &Parameter::getValue<double>)
-      .def("get_value_string", &Parameter::getValue<std::string>);
-
-  py::class_<Parameters>(m_config, "ConfigParameters")
-      .def(py::init<>())
-      .def("add_parameter",
-           (void(Parameters::*)(const std::string &, const Parameter &)) &
-               Parameters::addParameter)
-      .def("add_parameter", (void(Parameters::*)(const std::string &, bool)) &
-                                Parameters::addParameter)
-      .def("add_parameter",
-           (void(Parameters::*)(const std::string &, const char *)) &
-               Parameters::addParameter)
-      .def("add_parameter",
-           (void(Parameters::*)(const std::string &, int, int, int)) &
-               Parameters::addParameter)
-      .def("add_parameter",
-           (void(Parameters::*)(const std::string &, double, double, double)) &
-               Parameters::addParameter)
-      .def("from_dict", &set_parameters_from_dict);
-
-  // ------------------------------------------------------------------------------
-  // Control bindings submodule
+// Control bindings submodule
+void bindings_control(py::module_ &m) {
   auto m_control = m.def_submodule("control", "Control module");
 
   py::enum_<Control::ControlType>(m_control, "ControlType")
@@ -421,56 +237,5 @@ PYBIND11_MODULE(kompass_cpp, m) {
       .def("reset_target", &Control::VisionFollower::resetTarget)
       .def("get_ctrl", &Control::VisionFollower::getCtrl)
       .def("run", &Control::VisionFollower::run);
-
-  // ------------------------------------------------------------------------------
-  // Mapping bindings submodule
-  auto m_mapping = m.def_submodule("mapping", "Local Mapping module");
-  py::class_<Mapping::LocalMapper>(m_mapping, "LocalMapper")
-      .def(
-          py::init<int, int, float, const Eigen::Vector3f &, float, int, int>(),
-          py::arg("grid_height"), py::arg("grid_width"), py::arg("resolution"), py::arg("laserscan_position"),
-          py::arg("laserscan_orientation"), py::arg("max_points_per_line"), py::arg("max_num_threads") = 1)
-
-      .def(py::init<int, int, float, const Eigen::Vector3f &, float, float,
-                    float, float, float, float, float, int, int>(),
-           py::arg("grid_height"), py::arg("grid_width"), py::arg("resolution"), py::arg("laserscan_position"),
-           py::arg("laserscan_orientation"), py::arg("p_prior"), py::arg("p_empty"), py::arg("p_occupied"),
-           py::arg("range_sure"), py::arg("range_max"), py::arg("wall_size"), py::arg("max_points_per_line"),
-           py::arg("max_num_threads") = 1)
-
-      .def("scan_to_grid", &Mapping::LocalMapper::scanToGrid,
-           "Convert laser scan data to occupancy grid", py::arg("angles"),
-           py::arg("ranges"), py::arg("grid_data"))
-
-      .def("scan_to_grid_baysian", &Mapping::LocalMapper::scanToGridBaysian,
-           "Convert laser scan data to occupancy grid, with baysian update",
-           py::arg("angles"), py::arg("ranges"), py::arg("grid_data"),
-           py::arg("grid_data_prob"), py::arg("previous_grid_data_prob"))
-
-      .def("local_to_grid", &Mapping::LocalMapper::localToGrid,
-           "Convert a point from local coordinates frame of the grid to "
-           "grid indices",
-           py::arg("pose_target_in_central"))
-
-      .def("get_previous_grid_in_current_pose",
-           &Mapping::LocalMapper::getPreviousGridInCurrentPose,
-           py::arg("current_position_in_previous_pose"),
-           py::arg("current_orientation_in_previous_pose"),
-           py::arg("previous_grid_data"), py::arg("unknown_value"));
-
-  py::enum_<Mapping::OccupancyType>(m_mapping, "OCCUPANCY_TYPE")
-      .value("UNEXPLORED", Mapping::OccupancyType::UNEXPLORED)
-      .value("EMPTY", Mapping::OccupancyType::EMPTY)
-      .value("OCCUPIED", Mapping::OccupancyType::OCCUPIED);
-
-  // ------------------------------------------------------------------------------
-  // Utils bindings submodule
-  py::enum_<LogLevel>(m, "LogLevel")
-      .value("DEBUG", LogLevel::DEBUG)
-      .value("INFO", LogLevel::INFO)
-      .value("WARNING", LogLevel::WARNING)
-      .value("ERROR", LogLevel::ERROR)
-      .export_values();
-  m.def("set_log_level", &setLogLevel, "Set the log level");
-  m.def("set_log_file", &setLogFile, "Set the log file");
 }
+
