+++ src/kompass_cpp/kompass_cpp/src/bindings/bindings_mapping.cpp
@@ -1,0 +1,70 @@
+#include <pybind11/eigen.h>
+#include <pybind11/functional.h>
+#include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
+
+#include "mapping/local_mapper.h"
+
+#ifdef GPU
+#include "mapping/local_mapper_gpu.h"
+#endif
+
+namespace py = pybind11;
+using namespace Kompass;
+
+// Mapping bindings submodule
+void bindings_mapping(py::module_ &m) {
+  auto m_mapping = m.def_submodule("mapping", "Local Mapping module");
+  py::enum_<Mapping::OccupancyType>(m_mapping, "OCCUPANCY_TYPE")
+      .value("UNEXPLORED", Mapping::OccupancyType::UNEXPLORED)
+      .value("EMPTY", Mapping::OccupancyType::EMPTY)
+      .value("OCCUPIED", Mapping::OccupancyType::OCCUPIED);
+
+  py::class_<Mapping::LocalMapper>(m_mapping, "LocalMapper")
+      .def(
+          py::init<int, int, float, const Eigen::Vector3f &, float, int, int>(),
+          py::arg("grid_height"), py::arg("grid_width"), py::arg("resolution"),
+          py::arg("laserscan_position"), py::arg("laserscan_orientation"),
+          py::arg("max_points_per_line"), py::arg("max_num_threads") = 1)
+
+      .def(py::init<int, int, float, const Eigen::Vector3f &, float, float,
+                    float, float, float, float, float, int, int>(),
+           py::arg("grid_height"), py::arg("grid_width"), py::arg("resolution"),
+           py::arg("laserscan_position"), py::arg("laserscan_orientation"),
+           py::arg("p_prior"), py::arg("p_empty"), py::arg("p_occupied"),
+           py::arg("range_sure"), py::arg("range_max"), py::arg("wall_size"),
+           py::arg("max_points_per_line"), py::arg("max_num_threads") = 1)
+
+      .def("scan_to_grid", &Mapping::LocalMapper::scanToGrid,
+           "Convert laser scan data to occupancy grid", py::arg("angles"),
+           py::arg("ranges"), py::arg("grid_data"))
+
+      .def("scan_to_grid_baysian", &Mapping::LocalMapper::scanToGridBaysian,
+           "Convert laser scan data to occupancy grid, with baysian update",
+           py::arg("angles"), py::arg("ranges"), py::arg("grid_data"),
+           py::arg("grid_data_prob"), py::arg("previous_grid_data_prob"))
+
+      .def("local_to_grid", &Mapping::LocalMapper::localToGrid,
+           "Convert a point from local coordinates frame of the grid to "
+           "grid indices",
+           py::arg("pose_target_in_central"))
+
+      .def("get_previous_grid_in_current_pose",
+           &Mapping::LocalMapper::getPreviousGridInCurrentPose,
+           py::arg("current_position_in_previous_pose"),
+           py::arg("current_orientation_in_previous_pose"),
+           py::arg("previous_grid_data"), py::arg("unknown_value"));
+
+#ifdef GPU
+  py::class_<Mapping::LocalMapperGPU>(m_mapping, "LocalMapperGPU")
+      .def(
+          py::init<int, int, float, const Eigen::Vector3f &, float, int, int>(),
+          py::arg("grid_height"), py::arg("grid_width"), py::arg("resolution"),
+          py::arg("laserscan_position"), py::arg("laserscan_orientation"),
+          py::arg("scan_size"), py::arg("max_points_per_line") = 32)
+
+      .def("scan_to_grid", &Mapping::LocalMapper::scanToGrid,
+           "Convert laser scan data to occupancy grid", py::arg("angles"),
+           py::arg("ranges"), py::arg("grid_data"));
+#endif
+}
